Make THERMOSTAT_HOME an external requirement.
Usage then becomes:
$ THERMOSTAT_HOME=/path/to/thermostat/distribution/target/image ./util/install-plugin.sh

+++ util/install-plugin.sh
@@ -8,9 +8,11 @@ PLUGIN_NAME=${PWD##*/}
 PLUGIN_VERSION=1.99.12-SNAPSHOT
 PLUGIN_ZIP=thermostat-${PLUGIN_NAME}-distribution
 
-THERMOSTAT_HOME=INSERT-PATH-TO-THERMOSTAT-REPOSITORY-HERE
-THERMOSTAT_IMAGE=${THERMOSTAT_HOME}/distribution/target/image
-PLUGIN_HOME="${THERMOSTAT_HOME}"/distribution/target/image/plugins
+if [ -z ${THERMOSTAT_HOME} ]; then
+  echo 1>&2 "THERMOSTAT_HOME environment variable not set. Required for installation. It should point to your Thermostat installation."
+  exit 1
+fi
+PLUGIN_HOME="${THERMOSTAT_HOME}/plugins"
 
 rm -rfv ${PLUGIN_HOME}/${PLUGIN_NAME} 
-unzip -o ./distribution/target/${PLUGIN_ZIP}-${PLUGIN_VERSION}.zip -d ${THERMOSTAT_IMAGE}
+unzip -o ./distribution/target/${PLUGIN_ZIP}-${PLUGIN_VERSION}.zip -d ${THERMOSTAT_HOME}
